feature: Implement Login
Any user can log in by using email and password from Login page.
On succeed `log in` operation shows `It works` message,
otherwise differ messages according to the problem.

MVVM architecture pattern is used to implement the feature.
LoginInteractor.kt serves as an additional layer for calling
the server api, processing the result, and converting it to the final form.
LoginViewModel.kt exposes `_isLoading` state to provide
if the `log in` request is being executed, `_error` state
to provide an error message and `_result` state for a success message.

Add ResourcesRepository.kt to store `log in` result strings.